refactor: extract Zm async wrapper into its own module
Move the async ZoneManager adapter to `src/zm.rs`, keeping `main.rs`
focused on CLI parsing and orchestration.
